AMQCPP-205
Fixing warnings found on windows build

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@713591 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/decaf/util/concurrent/TimeUnit.cpp
@@ -111,7 +111,7 @@ void TimeUnit::sleep( long long timeout ) const {
 ////////////////////////////////////////////////////////////////////////////////
 void TimeUnit::timedWait( Synchronizable* obj, long long timeout ) const {
     if( timeout > 0 ) {
-        long ms = toMillis( timeout );
+        long long ms = toMillis( timeout );
         int ns = excessNanos( timeout, ms );
         obj->wait( (unsigned long)ms );
         // TODO - Only have a wait for Milliseconds currently.
